orkut-voice-agent/requirements.txt
# Voice Agent Framework
livekit==0.9.0
livekit-agents==0.9.0
livekit-agents-silero==0.9.0
python-dotenv==1.0.0

# AI/ML
cerebras-cloud-sdk==1.0.0
openai==1.3.0

# Speech Services
cartesia==0.5.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiohttp==3.9.1

# Development
black==23.12.0
flake8==6.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
